contacts also fix the DeltaChatController.callMethod again
