+++ src/renderer/components/helpers/ChatMethods.js
@@ -26,7 +26,7 @@ export function openBlockContactDialog (screenContext, selectedChat) {
     const contact = selectedChat.contacts[0]
     screenContext.openDialog('ConfirmationDialog', {
       message: tx('ask_block_contact'),
-      cb: yes => yes && callDcMethod('blockContact', contact.id)
+      cb: yes => yes && callDcMethod('contacts.blockContact', contact.id)
     })
   }
 }
